Move to 2.1, fix permissions
