+++ src/main/java/io/cryostat/net/web/http/api/beta/HttpApiBetaModule.java
@@ -39,7 +39,6 @@ package io.cryostat.net.web.http.api.beta;
 
 import io.cryostat.net.web.http.RequestHandler;
 import io.cryostat.net.web.http.api.beta.graph.GraphModule;
-
 import dagger.Binds;
 import dagger.Module;
 import dagger.multibindings.IntoSet;
@@ -91,9 +90,4 @@ public abstract class HttpApiBetaModule {
     @IntoSet
     abstract RequestHandler bindTargetRecordingMetadataLabelsPostBodyHandler(
             TargetRecordingMetadataLabelsPostBodyHandler handler);
-
-    @Binds
-    @IntoSet
-    abstract RequestHandler bindTargetCredentialKeysGetHandler(
-            TargetCredentialKeysGetHandler handler);
 }
