+++ src/main/java/io/cryostat/net/web/http/api/v2/HttpApiV2Module.java
@@ -184,4 +184,9 @@ public abstract class HttpApiV2Module {
     @Binds
     @IntoSet
     abstract RequestHandler bindReportGetHandler(ReportGetHandler handler);
+
+    @Binds
+    @IntoSet
+    abstract RequestHandler bindTargetCredentialKeysGetHandler(
+            TargetCredentialKeysGetHandler handler);
 }
